:hammer: refactor: Move image api to user domain
